Add tests for room creation and joining via websocket
